+++ frontend/src/screens/VerificationScreen.js
@@ -2,29 +2,38 @@ import React, {useEffect} from 'react';
 
 import {useDispatch, useSelector} from "react-redux";
 
-import '../css/verificationScreen.css'
+import {useHistory} from 'react-router-dom';
 
-import {verify} from '../actions/userActions.js'
+import '../css/verificationScreen.css';
 
-const VerificationScreen = (props, {history}) => {
+import {Alert, Spinner} from 'react-bootstrap';
+
+import {verify} from '../actions/userActions.js';
+
+const VerificationScreen = (props) => {
 
     const code = props.match.params.code;
 
     const dispatch = useDispatch();
+    const userRegister = useSelector(state => state.userRegister);
+    const {error} = userRegister;
+    let history = useHistory();
 
-    const registrationSuccess = () => {
+
+    useEffect(() => {
         dispatch(verify(code));
-    }
+        if(!error){
+            history.push("/admin");
+        }
+    }, [history]);
 
     return(
         <div className="body-verify">
-            <div className="block-message-verify">
-                <div className="verify-message">
-                    <button onClick={registrationSuccess}>
-                        Пройдите верификацию
-                    </button>
-                </div>
-            </div>
+            <Alert>
+                Подтверждение через почту
+                <Spinner animation="border"/>
+                {error}
+            </Alert>
         </div>
     )
 }